Move metamask redux focused ledger selectors to metamask duck
