+++ ui/selectors/selectors.js
@@ -40,6 +40,8 @@ import {
   getConversionRate,
   isNotEIP1559Network,
   isEIP1559Network,
+  getLedgerTransportType,
+  isAddresLedger,
 } from '../ducks/metamask/metamask';
 import { getLedgerWebHidConnectedStatus } from '../ducks/app/app';
 
@@ -75,7 +77,7 @@ export function getCurrentChainId(state) {
   return chainId;
 }
 
-function findKeyringForAddress(state, address) {
+export function findKeyringForAddress(state, address) {
   const simpleAddress = stripHexPrefix(address).toLowerCase();
 
   const keyring = state.metamask.keyrings.find((kr) => {
@@ -651,16 +653,6 @@ export function getTokenList(state) {
   return state.metamask.tokenList;
 }
 
-export function getLedgerTransportType(state) {
-  return state.metamask.ledgerTransportType;
-}
-
-export function isAddresLedger(state, address) {
-  const keyring = findKeyringForAddress(state, address);
-
-  return keyring?.type === KEYRING_TYPES.LEDGER;
-}
-
 export function doesAddressRequireLedgerHidConnection(state, address) {
   const addressIsLedger = isAddresLedger(state, address);
   const transportTypePreferenceIsWebHID =
@@ -672,10 +664,4 @@ export function doesAddressRequireLedgerHidConnection(state, address) {
   return (
     addressIsLedger && transportTypePreferenceIsWebHID && webHidIsNotConnected
   );
-}
-
-export function doesUserHaveALedgerAccount(state) {
-  return state.metamask.keyrings.some((kr) => {
-    return kr.type === KEYRING_TYPES.LEDGER;
-  });
 }
